pyproject.toml
[project]
name = "mcp-server-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-mcp-adapters>=0.1.10",
    "langchain-openai>=0.3.33",
    "langgraph>=0.6.7",
    "mcp[cli]>=1.15.0",
    "python-dotenv>=1.1.1",
]
